Coexp table - Sort by q value, bold significant
Also reorganize util function to formatUtils
