+++ src/pages/resultsView/enrichments/EnrichmentsUtil.spec.tsx
@@ -1,7 +1,7 @@
 import { assert } from 'chai';
 import React from 'react';
-import { calculateAlterationTendency, calculateExpressionTendency, formatLogOddsRatio, formatValueWithStyle, 
-    formatPercentage, getAlterationScatterData, getExpressionScatterData, roundLogRatio, 
+import { calculateAlterationTendency, calculateExpressionTendency,
+    formatPercentage, getAlterationScatterData, getExpressionScatterData,
     getAlterationRowData, getExpressionRowData, getFilteredData, getBarChartTooltipContent, getBoxPlotScatterData, 
     getDownloadContent, getAlterationsTooltipContent, shortenGenesLabel, getBoxPlotModels
 } from "./EnrichmentsUtil";
@@ -350,34 +350,6 @@ describe("EnrichmentsUtil", () => {
         });
     });
 
-    describe("#formatLogOddsRatio()", () => {
-        it("returns <-10 for -11", () => {
-            assert.equal(formatLogOddsRatio(-11), "<-10");
-        });
-
-        it("returns >10 for 11", () => {
-            assert.equal(formatLogOddsRatio(11), ">10");
-        });
-
-        it("returns 10 for 10", () => {
-            assert.equal(formatLogOddsRatio(10), "10.00");
-        });
-
-        it("returns 1.23 for 1.234", () => {
-            assert.equal(formatLogOddsRatio(1.234), "1.23");
-        });
-    });
-
-    describe("#formatValueWithStyle()", () => {
-        it("returns <span>0.300</span> for 0.3", () => {
-            expect(formatValueWithStyle(0.3)).toEqualJSX(<span>0.300</span>);
-        });
-
-        it("returns <b><span>0.030</span></b> for 0.03", () => {
-            expect(formatValueWithStyle(0.03)).toEqualJSX(<b><span>0.0300</span></b>);
-        });
-    });
-
     describe("#formatPercentage()", () => {
         it("returns 5 (10.35%) for 5 and 10.3452", () => {
             assert.equal(formatPercentage(5, 10.3452), "5 (10.35%)");
@@ -415,20 +387,6 @@ describe("EnrichmentsUtil", () => {
                 {x: 2.652285592481328, y: 8.249349711250797, hugoGeneSymbol: "EPHB3", entrezGeneId: 2049,
                     logRatio: 2.652285592481328, qValue: 0.000024395252515979897, hovered: false}
             ]);
-        });
-    });
-
-    describe("#roundLogRatio()", () => {
-        it("returns 5 for 8 and 5", () => {
-            assert.equal(roundLogRatio(8, 5), 5);
-        });
-
-        it("returns -3 for -4 and 3", () => {
-            assert.equal(roundLogRatio(-4, 3), -3);
-        });
-
-        it("returns 3.21 for 3.2123 and 10", () => {
-            assert.equal(roundLogRatio(3.2123, 10), 3.21);
         });
     });
